+++ src/reducers/profileReducer.js
@@ -1,8 +1,23 @@
-import React from 'react'
-
-const initialState = {}
-const profileReducer = (state = initialState, action) => {
-  
+const initialState = {
+  loggedInAllegiances: [],
+  error: [],
 }
 
-export default profileReducer
+export const profileActions = (state = initialState, action) => {
+  switch (action.type) {
+    case types.FETCH_PROFILE_SUCCESS:
+      //Refreshes logged in user's info, groups and allegiances upon entering their profile.
+      return {
+        ...state,
+        loggedInAllegiances: action.payload.basicAllegianceInfo,
+        error: '',
+      }
+    case profileTypes.FETCH_PROFILE_FAILURE:
+      return {
+        ...state,
+        error: action.payload,
+      }
+    default:
+      return state
+  }
+}
